Add some more command line options
The command line options title and length are added by reusing the logic
implemented in info, but allowing the user only to ask specifically for a
single information.

In order to navigate in the track, the position and seek commands are added
both in the Player as well as in the ClientSubcommands class.