Catch special exceptions that occure if no VLC instance is running
In case a command is sent to class Player and VLC is not running, the following
error occurs:
        <<'NoneType' object has no attribute 'xyz'.>>
This error is explicitly catched in player_list_error_wrapped()  to avoid
crashing of the complete program.

All methods that rely on a valid response of their called function need a
special error handling.