Add a fallback for streams with no entry in 'title'
In case a stream is played as a raw media file with bad metadate, it might
happen that no valid content is return by the DBUS command for the title. In
this case a fallback is added to display the path to the file instead. The path
is always a valid content.